Add basic tests for kafka client
