+++ events_processor/config/kafka/kafka_test.go
@@ -1,0 +1,1 @@
+package kafka
